boot/bootutil: Split private image API out of image.h
image.h will now only contain applicaiton image related structures
and defines.
The split has been done to reduce need for including extra headers,
for example FIH support and mcuboot_config.h into software that
needs the same defining as MCUboot but implements own functions.
